content/svg/content/test/test_onerror.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=500261
-->
<head>
  <title>Test onerror behaviour</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=500261">Mozilla Bug 500261</a>
<p id="display"></p>
<script class="testbody" type="text/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function run()
{
  ok(true, 'onerror method called');
  SimpleTest.finish();
}

]]>
</script>
<div id="content">

  <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="100%" height="1" id="svg">
    <image width="1" height="1" xlink:href="http://localhost/serverGone.gif" onerror="run()"/>
  </svg>

</div>
<pre id="test">
</pre>
</body>
</html>
